refactor: move more read models into admin ui module
