+++ src/modules/reporting/index.ts
@@ -7,23 +7,3 @@ export {
   type BookProgressRow,
   type BookProgressContributor,
 } from "./read-models/getLanguageBookProgressReadModel";
-export {
-  getLanguageDashboardBooksReadModel,
-  type LanguageDashboardBookReadModel,
-} from "./read-models/getLanguageDashboardBooksReadModel";
-export {
-  getLanguageDashboardMembersReadModel,
-  type LanguageDashboardMemberReadModel,
-} from "./read-models/getLanguageDashboardMembersReadModel";
-export {
-  getLanguageDashboardContributionsReadModel,
-  type LanguageDashboardContributionReadModel,
-} from "./read-models/getLanguageDashboardContributionsReadModel";
-export {
-  getLanguageDashboardActivityReadModel,
-  type LanguageDashboardActivityEntryReadModel,
-} from "./read-models/getLanguageDashboardActivityReadModel";
-export {
-  getLanguageApprovalActivityReadModel,
-  type LanguageApprovalActivityReadModel,
-} from "./read-models/getLanguageApprovalActivityReadModel";
