+++ src/ui/admin/readModels/getAIGlossImportLanguagesReadModel.ts
@@ -1,4 +1,4 @@
-import { aiGlossImportService } from "../data-access/aiGlossImportService";
+import { aiGlossImportService } from "@/modules/translation/data-access/aiGlossImportService";
 
 // TODO: Cache this
 export function getAIGlossImportLanguagesReadModel() {
